hfi1_eprom: Fix filenames when generating multiple files
Commit 43445f0b0936 ("hfi1_eprom: separate files for collective
reads" ) introduced a change to handle repeating filenames when
dealing with systems that have multiple devices.

However, that change is broken in instances when the tool is
writing multiple files for the same device. Consequently, the
command 'hfi1_eprom -c config -b bulk -o op -r' would end up
producing files named 'config', 'bulk_1' and 'op_2' despite the
fact that the tool might be operating on a singe device.

This was due to the fact that the change above used a single
global variable that was being incremented for every file generated
without checking whether the file existed first.

The fix is to first check for the existence of the target file. If
it does exist, use the next index up. This handles not only operating
on multiple devices but also overwriting files from previous
invocations of the tool.

Fixes: 43445f0b0936 ("hfi1_eprom: separate files for collective reads")
Reviewed-by: John Fleck <john.fleck@intel.com>
Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>